Split pid again
